+++ src/Data/DAL/INRONotificationRepository.cs
@@ -12,7 +12,7 @@ namespace Prizm.Data.DAL
     {
         List<NotRequiredOperation> GetAllNotRequiredOperation();
         List<KeyValuePair<DateTime, Guid>> GetAllNotRequiredOperationResult();
-        KeyValuePair<Guid, float> GetAllUnitsProducedSinceLastDate(Guid testId, DateTime maxDate, FrequencyMeasure measure);
-        KeyValuePair<Guid, float> GetUnitsProducedSinceLastDateTest(Guid testId, FrequencyMeasure measure);
+        KeyValuePair<Guid, decimal> GetAllUnitsProducedSinceLastDate(Guid testId, DateTime maxDate, FrequencyMeasure measure);
+        KeyValuePair<Guid, decimal> GetUnitsProducedSinceLastDateTest(Guid testId, FrequencyMeasure measure);
     }
 }
